requirements.txt
opencv-python-headless
numpy
joblib
scikit-learn
pillow
streamlit
gdown
collections